requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
diff-match-patch==20121119
Django==1.11.11
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-filter==2.0.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.1.0
django-js-asset==1.1.0
django-redis==4.9.0
django-reversion==3.0.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
drf-haystack==1.8.4
elasticsearch==2.4.1
et-xmlfile==1.0.1
fdfs-client-py==1.2.6
future==0.17.1
httplib2==0.9.2
idna==2.7
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
mutagen==1.41.1
odfpy==1.3.6
openpyxl==2.5.9
Pillow==5.3.0
pycparser==2.19
pycryptodomex==3.7.0
PyJWT==1.6.4
PyMySQL==0.9.2
python-alipay-sdk==1.8.0
python-dateutil==2.7.5
pytz==2018.6
PyYAML==3.13
QQLoginTool==0.3.0
redis==2.10.6
requests==2.20.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24
uWSGI==2.0.17.1
vine==1.1.4
xadmin==0.6.1
xlrd==1.1.0
xlwt==1.3.0
